+++ src/veins/modules/application/platooning/protocols/BaseProtocol.cc
@@ -55,9 +55,9 @@ void BaseProtocol::initialize(int stage)
         // beaconing interval in seconds
         beaconingInterval = SimTime(par("beaconingInterval").doubleValue());
         // platooning message packet size
-        packetSize = par("packetSize").longValue();
+        packetSize = par("packetSize").intValue();
         // priority of platooning message
-        priority = par("priority").longValue();
+        priority = par("priority").intValue();
         ASSERT2(priority >= 0 && priority <= 7, "priority value must be between 0 and 7");
 
         // init messages for scheduleAt
